Revert "Code optimisation, inheritance fixes"
