+++ Assets/Scripts/CustomTankCreator.cs
@@ -1,27 +1,18 @@
-using System;
 using ExitGames.Client.Photon;
 using Photon.Pun;
-using Photon.Realtime;
 using UnityEngine;
 using Random = UnityEngine.Random;
-using TMPro;
 
-public class CustomTankCreator : MonoBehaviourPunCallbacks, IPunObservable
+public class CustomTankCreator : MonoBehaviourPunCallbacks
 {
-    [Header("Visuals")]
-    [SerializeField] private TMP_Text nickname;
-    [Space(5)]
-    [SerializeField] private SpriteRenderer tankBase;
-    [SerializeField] private SpriteRenderer tankTower;
-    [Space(10)]
-    [Header("Random")]
-    [SerializeField] private AnimatorOverrideController[] tankBases;
-    [SerializeField] private Sprite[] tankTowers;
-    public Player Player { get; private set; }
+    [SerializeField] protected SpriteRenderer tankBase;
+    [SerializeField] protected SpriteRenderer tankTower;
+    [SerializeField] protected Animator tankBaseAnimator;
+    [SerializeField] protected AnimatorOverrideController[] tankBases;
+    [SerializeField] protected Sprite[] tankTowers;
 
     private float r, g, b;
-    private PhotonView view;
-    private Animator tankBaseAnimator;
+    protected PhotonView view;
 
     private string randomBase, randomTower;
 
@@ -38,7 +29,6 @@ public class CustomTankCreator : MonoBehaviourPunCallbacks, IPunObservable
     public override void OnEnable()
     {
         view = GetComponent<PhotonView>();
-        tankBaseAnimator = GetComponent<Animator>();
         
         PhotonNetwork.AddCallbackTarget(this);
         
@@ -67,59 +57,5 @@ public class CustomTankCreator : MonoBehaviourPunCallbacks, IPunObservable
         if(!view.IsMine) return;
         PhotonNetwork.SetPlayerCustomProperties(_customProperties);
         //view.RPC("SetTexturesAndColor", RpcTarget.All, r, g, b, randomBase, randomTower);
-    }
-    
-     private void SetPlayerInfo(Player player)
-    {
-        Player = player;
-        SetTextures(player);
-    }
-
-    public override void OnPlayerPropertiesUpdate(Player target, ExitGames.Client.Photon.Hashtable changedProps)
-    {
-        base.OnPlayerPropertiesUpdate(target,changedProps);
-        if (target != null || target == Player)
-        {
-            if(changedProps.ContainsKey("R") && changedProps.ContainsKey("G") && changedProps.ContainsKey("B")
-               && changedProps.ContainsKey("tankBase") && changedProps.ContainsKey("tankTower"))
-                SetTextures(target);
-        }
-    }
-    
-    private void SetTextures(Player player)
-    {
-        float r = 0;
-        float g = 0;
-        float b = 0;
-        string _tankBase = null;
-        string _tankTower = null;
-
-        if (player.CustomProperties.ContainsKey("R"))
-            r = (float) player.CustomProperties["R"];
-        if (player.CustomProperties.ContainsKey("G"))
-            g = (float) player.CustomProperties["G"];
-        if (player.CustomProperties.ContainsKey("B"))
-            b = (float) player.CustomProperties["B"];
-        
-        if (player.CustomProperties.ContainsKey("tankBase"))
-            _tankBase = (string) player.CustomProperties["tankBase"];
-        if (player.CustomProperties.ContainsKey("tankTower"))
-            _tankTower = (string) player.CustomProperties["tankTower"];
-
-        if (player.CustomProperties.ContainsKey("nickname"))
-            nickname.text = (string) player.CustomProperties["nickname"];
-        
-        var randColor = new Color(r, g, b);
-        tankBase.color = randColor;
-        tankBaseAnimator.runtimeAnimatorController = Array.Find(tankBases, s => s.name == _tankBase);
-        
-        tankTower.color = randColor;
-        tankTower.sprite = Array.Find(tankTowers, s => s.name == _tankTower);
-    }
-
-    public void OnPhotonSerializeView(PhotonStream stream, PhotonMessageInfo info)
-    {
-        info.Sender.TagObject = gameObject;
-        SetPlayerInfo(info.Sender);
     }
 }
