+++ Assets/Scripts/TankSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4406e7ca3505ee849a333333bad0253a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
